use std::sqrt instead of sqrtf

+++ include/insns/RV32F.h
@@ -48,7 +48,7 @@ class RV32F : public RevExt{
   static constexpr auto& fcvtwus = CvtFpToInt<float, uint32_t>;
 
   static bool fsqrts(RevFeature *F, RevRegFile *R, RevMem *M, const RevInst& Inst) {
-    R->SetFP(Inst.rd, sqrtf( R->GetFP<float>(Inst.rs1) ));
+    R->SetFP(Inst.rd, std::sqrt( R->GetFP<float>(Inst.rs1) ));
     R->AdvancePC(Inst);
     return true;
   }
